requirements.txt
pandas
jinja2
matplotlib
openpyxl
python-dotenv
apscheduler
requests
fastapi
uvicorn[standard]
plotly
weasyprint
apscheduler>=3.10
